Finalize restructuring for chunked operations
